Reorganize package using include folder.
