+++ rosserial_python/nodes/serial_node.py
@@ -37,10 +37,12 @@ __author__ = "mferguson@willowgarage.com (Michael Ferguson)"
 
 import rospy
 from rosserial_python import SerialClient, RosSerialServer
+from serial import SerialException
+from time import sleep
 import multiprocessing
 
 import sys
-   
+
 if __name__=="__main__":
 
     rospy.init_node("serial_node")
@@ -59,7 +61,7 @@ if __name__=="__main__":
         port_name  = sys.argv[1]
     if len(sys.argv) == 3 :
         tcp_portnum = int(sys.argv[2])
-    
+
     if port_name == "tcp" :
         server = RosSerialServer(tcp_portnum, fork_server)
         rospy.loginfo("Waiting for socket connections on port %d" % tcp_portnum)
@@ -75,11 +77,22 @@ if __name__=="__main__":
                 process.join()
             rospy.loginfo("All done")
 
-    else :          # Use serial port 
-        rospy.loginfo("Connecting to %s at %d baud" % (port_name,baud) )
-        client = SerialClient(port_name, baud)
-        try:
-            client.run()
-        except KeyboardInterrupt:
-            pass
+    else :          # Use serial port
+        while not rospy.is_shutdown():
+            rospy.loginfo("Connecting to %s at %d baud" % (port_name,baud) )
+            try:
+                rospy.loginfo("qqq")
+                client = SerialClient(port_name, baud)
+                rospy.loginfo("aaa")
+                client.run()
+            except KeyboardInterrupt:
+                break
+            except SerialException:
+                rospy.loginfo("xxxx")
+                sleep(1.0)
+                continue
+            except OSError:
+                rospy.loginfo("zzzz")
+                sleep(1.0)
+                continue
