+++ rosserial_test/test/rosserial_python_serial.test
@@ -1,8 +1,9 @@
 <launch>
-  <node pkg="rosserial_python" type="serial_node.py" name="rosserial_python_node" args="/tmp/rosserial_pty" />
-  <test test-name="rosserial_server_socket_test_publish_subscribe" pkg="rosserial_test"
+  <node pkg="rosserial_python" type="serial_node.py" name="rosserial_python_node" args="/tmp/rosserial_pty2" output="screen" />
+
+  <test test-name="rosserial_python_serial_test_publish_subscribe" pkg="rosserial_test"
         type="rosserial_test_publish_subscribe" time-limit="10.0">
     <param name="mode" value="serial" />
-    <param name="port" value="/tmp/rosserial_pty" />
+    <param name="port" value="/tmp/rosserial_pty2" />
   </test>
 </launch>
