+++ rosserial_test/test/rosserial_python_socket.test
@@ -1,5 +1,6 @@
 <launch>
   <node pkg="rosserial_python" type="serial_node.py" name="rosserial_python_node" args="tcp" />
+
   <test test-name="rosserial_server_socket_test_publish_subscribe" pkg="rosserial_test"
         type="rosserial_test_publish_subscribe" time-limit="10.0">
     <param name="mode" value="socket" />
